Connects the pieces together. There's one other TODO and then it will work for checkstyle :)
Now we can use a gradle task that will clone the repo in a subfolder, run the analisys in the default branch, find the analysis from the current branch, compare them and print the results in the terminal. The last bit is to compare the same files that are in different paths.
This adds one more field to the Issue class (column), fixes the parsing when a tag is empty